refactor: 요소 컴퍼넌트화 #9
