+++ src/popUpPage/components/Background.tsx
@@ -1,0 +1,5 @@
+export default function Background() {
+  return (
+    <div className="dark:bg-donutool-bg absolute -z-50 h-[400px] w-[350px] bg-gray-300 transition duration-300"></div>
+  );
+}
