+++ src/popUpPage/components/Title.tsx
@@ -1,0 +1,7 @@
+export default function Title() {
+  return (
+    <h2 className="dark:text-donutool-text mb-4 text-3xl font-black text-neutral-600 transition duration-300 select-none">
+      DonuTool
+    </h2>
+  );
+}
